Integer -> Long

+++ persistence/spring-data-jpa/src/main/java/lobster/persistence/model/Activity.java
@@ -22,7 +22,7 @@ public class Activity implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Getter
     @Setter
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Getter
@@ -31,11 +31,11 @@ public class Activity implements Serializable {
 
     @Getter
     @Setter
-    private Integer calories;
+    private Long calories;
 
     @Getter
     @Setter
-    private Integer fatLevel;
+    private Long fatLevel;
 
     @Getter
     @Setter
@@ -47,10 +47,10 @@ public class Activity implements Serializable {
 
     @Getter
     @Setter
-    private Integer hours;
+    private Long hours;
 
     @Getter
     @Setter
-    private Integer happiness;
+    private Long happiness;
 
 }
